+++ Dev/ObjectiveCAppExample/main.m
@@ -1,0 +1,18 @@
+//
+//  main.m
+//  ObjectiveCAppExample
+//
+//  Created by Helen Jiang on 1/2/24.
+//
+
+#import <UIKit/UIKit.h>
+#import "AppDelegate.h"
+
+int main(int argc, char * argv[]) {
+    NSString * appDelegateClassName;
+    @autoreleasepool {
+        // Setup code that might create autoreleased objects goes here.
+        appDelegateClassName = NSStringFromClass([AppDelegate class]);
+    }
+    return UIApplicationMain(argc, argv, nil, appDelegateClassName);
+}